Cardano.CLI

Description

Shared types and helpers for CLI parsing

CLI Execution

cli :: Mod CommandFields a -> ParserInfo a

Construct a CLI from a list of a commands

>>> runCli $ cli $ cmdA <> cmdB <> cmdC

runCli :: ParserInfo (IO ()) -> IO ()

Runs a specific command parser using appropriate preferences

Commands

cmdMnemonic :: Mod CommandFields (IO ())

cmdWallet :: (ToJSON wallet, CmdWalletUpdatePassphrase wallet) => CmdWalletCreate wallet -> WalletClient wallet -> Mod CommandFields (IO ())

cmdWalletCreate :: WalletClient ApiWallet -> Mod CommandFields (IO ())

cmdByronWalletCreate :: WalletClient ApiByronWallet -> Mod CommandFields (IO ())

cmdTransaction :: ToJSON wallet => TransactionClient -> WalletClient wallet -> Mod CommandFields (IO ())

cardano-wallet transaction

cmdAddress :: AddressClient -> Mod CommandFields (IO ())

cmdStakePool :: ToJSON apiPool => StakePoolClient apiPool -> Mod CommandFields (IO ())

cmdNetwork :: NetworkClient -> Mod CommandFields (IO ())

cmdVersion :: Mod CommandFields (IO ())

cmdKey :: Mod CommandFields (IO ())

Option & Argument Parsers

optionT :: FromText a => Mod OptionFields a -> Parser a

Helper for writing an option Parser using a FromText instance.

argumentT :: FromText a => Mod ArgumentFields a -> Parser a

Helper for writing an argument Parser using a FromText instance.

databaseOption :: Parser FilePath

-database=DIR

hostPreferenceOption :: Parser HostPreference

--listen-address=HOSTSPEC

, default: 127.0.0.1

listenOption :: Parser Listen

--random-port|--port=INT

shutdownHandlerFlag :: Parser Bool

--shutdown-handler

stateDirOption :: FilePath -> Parser (Maybe FilePath)

-state-dir=DIR, default: ~.cardano-wallet$backend/$network

syncToleranceOption :: Parser SyncTolerance

-sync-tolerance=DURATION, default: 300s

tlsOption :: Parser TlsConfiguration

poolMetadataSourceOption :: Parser PoolMetadataSource

tokenMetadataSourceOption :: Parser TokenMetadataServer

--token-metadata-server=URL

metadataOption :: Parser (Maybe TxMetadataWithSchema)

--metadata=JSON

Note: we decode the JSON just so that we can validate more client-side.

timeToLiveOption :: Parser (Maybe (Quantity "second" NominalDiffTime))

--ttl=DURATION

Option parsers for configuring tracing

data LoggingOptions tracers

Constructors

LoggingOptions

Fields

loggingMinSeverity :: Severity

loggingTracers :: tracers

loggingTracersDoc :: Maybe Void

data LogOutput

Constructors

LogToStdStreams Severity
Log to console, with the given minimum Severity.

Logs of Warning or higher severity will be output to stderr. Notice or lower severity logs will be output to stdout.

LogToFile FilePath Severity

helperTracing :: [(String, String)] -> Parser (a -> a)

A hidden "helper" option which always fails, but shows info about the logging options.

loggingOptions :: Parser tracers -> Parser (LoggingOptions tracers)

loggingSeverities :: [(String, Severity)]

The lower-case names of all Severity values.

parseLoggingSeverity :: String -> Either String Severity

loggingSeverityOrOffReader :: ReadM (Maybe Severity)

loggingSeverityReader :: ReadM Severity

Types

data Service

Wrapper type around Text to make its semantic more explicit

data TxId

newtype Port (tag :: Symbol)

Port number with a tag for describing what it is used for

Constructors

Port

Fields

getPort :: Int

Logging

withLogging :: [LogOutput] -> ((Switchboard Text, (Configuration, Trace IO Text)) -> IO a) -> IO a

See withLoggingNamed

withLoggingNamed

Arguments

:: LoggerName

-> [LogOutput]

-> ((Switchboard Text, (Configuration, Trace IO Text)) -> IO a)
The action to run with logging configured.

-> IO a

Run an action with logging available and configured. When the action is finished (normally or otherwise), log messages are flushed.

ANSI Terminal Helpers

putErrLn :: Text -> IO ()

Like hPutErrLn but with provided default Handle

hPutErrLn :: Handle -> Text -> IO ()

Print an error message in red

enableWindowsANSI :: IO ()

The IOHK logging framework prints out ANSI colour codes with its messages. On Windows 10 and above it's possible to enable processing of these colour codes. The hSupportsANSIWithoutEmulation function does this as a side effect. On older versions of Windows, special treatment is required (see: ANSI). In this case, this function will achieve nothing, and the ANSI control characters will be printed in grey (too bad).

Working with Sensitive Data

getLine :: Text -> (Text -> Either String a) -> IO (a, Text)

Like hGetLine but with default handles

hGetLine :: Buildable e => (Handle, Handle) -> Text -> (Text -> Either e a) -> IO (a, Text)

Prompt user and parse the input. Re-prompt on invalid inputs.

getSensitiveLine

Arguments

:: Buildable e

=> Text
A message to prompt the user

-> (Text -> Either e a)
An explicit parser from Text

-> IO (a, Text)

Like hGetSensitiveLine but with default handles

hGetSensitiveLine :: Buildable e => (Handle, Handle) -> Text -> (Text -> Either e a) -> IO (a, Text)

Gather user inputs until a newline is met, hiding what's typed with a placeholder character.

Helpers

decodeError :: ByteString -> Maybe Text

Decode API error messages and extract the corresponding message.

requireFilePath :: FilePath -> IO ()

Look whether a particular filepath is correctly resolved on the filesystem. This makes for a better user experience when passing wrong filepaths via options or arguments, especially when they get forwarded to other services.

getDataDir

Arguments

:: String
The network backend name.

-> IO FilePath

Find the user data directory for a given node network backend.

setupDirectory :: (Text -> IO ()) -> FilePath -> IO ()

Initialize a directory to store data such as blocks or the wallet databases

getPrometheusURL :: IO (Maybe (String, Port "Prometheus"))

getEKGURL :: IO (Maybe (String, Port "EKG"))

ekgEnabled :: IO Bool