Cardano.CLI

Description

Shared types and helpers for CLI parsing

CLI Execution

cli :: Mod CommandFields a -> ParserInfo a

Construct a CLI from a list of a commands

>>> runCli $ cli $ cmdA <> cmdB <> cmdC

runCli :: ParserInfo (IO ()) -> IO ()

Runs a specific command parser using appropriate preferences

Commands

cmdMnemonic :: Mod CommandFields (IO ())

cmdWallet :: forall t. (DecodeAddress t, EncodeAddress t) => Mod CommandFields (IO ())

cmdTransaction :: forall t. (DecodeAddress t, EncodeAddress t) => Mod CommandFields (IO ())

cardano-wallet transaction

cmdAddress :: forall t. (DecodeAddress t, EncodeAddress t) => Mod CommandFields (IO ())

cmdVersion :: Mod CommandFields (IO ())

execLaunch :: Verbosity -> FilePath -> (Trace IO Text -> FilePath -> IO ()) -> [Command] -> IO ()

Execute launch commands. This differs from the serve command as it takes care of also starting a node backend in two separate processes and monitors both processes: if one terminates, then the other one is cancelled.

Option & Argument Parsers

optionT :: FromText a => Mod OptionFields a -> Parser a

Helper for writing an option Parser using a FromText instance.

argumentT :: FromText a => Mod ArgumentFields a -> Parser a

Helper for writing an argument Parser using a FromText instance.

databaseOption :: Parser FilePath

-database=FILEPATH

listenOption :: Parser Listen

--random-port|--port=INT

nodePortOption :: Parser (Port "Node")

--node-port=INT

, default: 8080

stateDirOption :: FilePath -> Parser (Maybe FilePath)

-state-dir=FILEPATH, default: ~.cardano-wallet$backend/$network

verbosityOption :: Parser Verbosity

(--quiet|--verbose)

Types

data Service

Wrapper type around Text to make its semantic more explicit

data MnemonicSize

Represents the number of words in a mnemonic sentence.

Only valid sizes are representable by this type.

Constructors

MS_9

MS_12

MS_15

MS_18

MS_21

MS_24

newtype Port (tag :: Symbol)

Port number with a tag for describing what it is used for

Constructors

Port

Fields

getPort :: Int

Logging

data Verbosity

Controls how much information to include in log output.

Constructors

Default
The default level of verbosity.

Quiet
Include less information in the log output.

Verbose
Include more information in the log output.

initTracer :: Severity -> Text -> IO (Configuration, Switchboard Text, Trace IO Text)

Initialize logging at the specified minimum Severity level.

verbosityToArgs :: Verbosity -> [String]

Convert a given Verbosity level into a list of command line arguments that can be passed through to a sub-process.

verbosityToMinSeverity :: Verbosity -> Severity

Map a given Verbosity level onto a Severity level.

Unicode Terminal Helpers

setUtf8Encoding :: IO ()

Override the system output encoding setting. This is needed because the CLI prints UTF-8 characters regardless of the LANG environment variable.

ANSI Terminal Helpers

putErrLn :: Text -> IO ()

Like hPutErrLn but with provided default Handle

hPutErrLn :: Handle -> Text -> IO ()

Print an error message in red

Working with Sensitive Data

getLine :: Buildable e => Text -> (Text -> Either e a) -> IO (a, Text)

Like hGetLine but with default handles

hGetLine :: Buildable e => (Handle, Handle) -> Text -> (Text -> Either e a) -> IO (a, Text)

Prompt user and parse the input. Re-prompt on invalid inputs.

getSensitiveLine

Arguments

:: Buildable e

=> Text
A message to prompt the user

-> (Text -> Either e a)
An explicit parser from Text

-> IO (a, Text)

Like hGetSensitiveLine but with default handles

hGetSensitiveLine :: Buildable e => (Handle, Handle) -> Text -> (Text -> Either e a) -> IO (a, Text)

Gather user inputs until a newline is met, hiding what's typed with a placeholder character.

Helpers

decodeError :: ByteString -> Maybe Text

Decode API error messages and extract the corresponding message.

requireFilePath :: FilePath -> IO ()

Look whether a particular filepath is correctly resolved on the filesystem. This makes for a better user experience when passing wrong filepaths via options or arguments, especially when they get forwarded to other services.

getDataDir

Arguments

:: String
The network backend name.

-> IO FilePath

Find the user data directory for a given node network backend.

waitForService

Arguments

:: Service
Name of the service

-> (Switchboard Text, Trace IO Text)
A Trace for logging

-> Port "Node"
Underlying TCP port

-> IO ()
Service we're waiting after.

-> IO ()

Wait for a service to become available on a given TCP port. Exit on failure with a proper error message.