Cardano.CLI

Description

Shared types and helpers for CLI parsing

Environment

data Environment

A convenient representation of a complete CLI environment.

Constructors

Environment

Fields

args :: Arguments

cli :: Docopt

argPresent :: Option -> Bool

parseAllArgs :: forall a. FromText a => Option -> IO (NonEmpty a)

parseArg :: forall a. FromText a => Option -> IO a

parseOptionalArg :: forall a. FromText a => Option -> IO (Maybe a)

CLI Construction

makeCli :: String -> String -> String -> Docopt

Make a specialized CLI from the specified commands, options, and examples.

CLI Execution

runCli :: (Manager -> Environment -> IO ()) -> Docopt -> IO ()

Runs a CLI with the specified backend and CLI definition.

runCliCommand

Arguments

:: (DecodeAddress t, EncodeAddress t)

=> Environment

-> Manager

-> (Proxy t -> IO ())
execServe

-> IO ()
execLaunch

-> IO ()

Interprets the given environment and executes a CLI command.

Launch Support

execLaunchCommands :: Trace IO Text -> Maybe FilePath -> [Command] -> IO ()

Execute launch commands. This differs from the serve command as it takes care of also starting a node backend in two separate processes and monitors both processes: if one terminates, then the other one is cancelled.

commandWalletServe :: String -> Listen -> Port "Node" -> Maybe FilePath -> String -> Verbosity -> Maybe (Hash "Genesis") -> Command

Creates a command that can be used to start the wallet API server.

parseWalletListen :: Environment -> IO Listen

Parse and convert the `--port` or `--random-port` option into a Listen data-type.

setupStateDir :: Trace IO Text -> FilePath -> IO ()

Initialize a state directory to store blockchain data such as blocks or the wallet database.

Types

newtype Port (tag :: Symbol)

Port number with a tag for describing what it is used for

Constructors

Port

Fields

getPort :: Int

Logging

data Verbosity

Controls how much information to include in log output.

Constructors

Default
The default level of verbosity.

Quiet
Include less information in the log output.

Verbose
Include more information in the log output.

initTracer :: Severity -> Text -> IO (Configuration, Trace IO Text)

Initialize logging at the specified minimum Severity level.

minSeverityFromArgs :: Arguments -> Severity

Determine the minimum Severity level from the specified command line arguments.

verbosityFromArgs :: Arguments -> Verbosity

Determine the desired Verbosity level from the specified command line arguments.

verbosityToArgs :: Verbosity -> [String]

Convert a given Verbosity level into a list of command line arguments that can be passed through to a sub-process.

verbosityToMinSeverity :: Verbosity -> Severity

Map a given Verbosity level onto a Severity level.

Mnemonics

execGenerateMnemonic :: Text -> IO ()

Generate a random mnemonic of the given size n (n = number of words), and print it to stdout.

Unicode Terminal Helpers

setUtf8Encoding :: IO ()

Override the system output encoding setting. This is needed because the CLI prints UTF-8 characters regardless of the LANG environment variable.

ANSI Terminal Helpers

putErrLn :: Text -> IO ()

Like hPutErrLn but with provided default Handle

hPutErrLn :: Handle -> Text -> IO ()

Print an error message in red

Parsing Arguments

newtype OptionValue a

A wrapper to avoid orphan instances for types defined externally.

Constructors

OptionValue

Fields

getOptionValue :: a

optional :: (Monoid m, Eq m) => (m -> Either e a) -> m -> Either e (Maybe a)

Make an existing parser optional. Returns 'Right Nothing' if the input is empty, without running the parser.

parseArgWith :: FromText a => Docopt -> Arguments -> Option -> IO a

parseAllArgsWith :: FromText a => Docopt -> Arguments -> Option -> IO (NonEmpty a)

help :: Docopt -> IO ()

Like exitWithUsage, but with a success exit code

Working with Sensitive Data

getLine :: Buildable e => Text -> (Text -> Either e a) -> IO (a, Text)

Like hGetLine but with default handles

hGetLine :: Buildable e => (Handle, Handle) -> Text -> (Text -> Either e a) -> IO (a, Text)

Prompt user and parse the input. Re-prompt on invalid inputs.

getSensitiveLine

Arguments

:: Buildable e

=> Text
A message to prompt the user

-> (Text -> Either e a)
An explicit parser from Text

-> IO (a, Text)

Like hGetSensitiveLine but with default handles

hGetSensitiveLine :: Buildable e => (Handle, Handle) -> Text -> (Text -> Either e a) -> IO (a, Text)

Gather user inputs until a newline is met, hiding what's typed with a placeholder character.

Helpers

decodeError :: ByteString -> Maybe Text

Decode API error messages and extract the corresponding message.

showT :: ToText a => a -> String

Show a data-type through its ToText instance