Cardano.CLI

Description

Shared types and helpers for CLI parsing

Types

newtype Port (tag :: Symbol)

Port number with a tag for describing what it is used for

Constructors

Port Int

Unicode Terminal Helpers

setUtf8Encoding :: IO ()

Override the system output encoding setting. This is needed because the CLI prints UTF-8 characters regardless of the LANG environment variable.

ANSI Terminal Helpers

putErrLn :: Text -> IO ()

Like hPutErrLn but with provided default Handle

hPutErrLn :: Handle -> Text -> IO ()

Print an error message in red

Parsing Arguments

parseArgWith :: FromText a => Docopt -> Arguments -> Option -> IO a

parseAllArgsWith :: FromText a => Docopt -> Arguments -> Option -> IO (NonEmpty a)

parseOptionalArg :: FromText a => Arguments -> Option -> IO (Maybe a)

help :: Docopt -> IO ()

Like exitWithUsage, but with a success exit code

Working with Sensitive Data

getLine :: Buildable e => Text -> (Text -> Either e a) -> IO (a, Text)

Like hGetLine but with default handles

hGetLine :: Buildable e => (Handle, Handle) -> Text -> (Text -> Either e a) -> IO (a, Text)

Prompt user and parse the input. Re-prompt on invalid inputs.

getSensitiveLine

Arguments

:: Buildable e

=> Text
A message to prompt the user

-> (Text -> Either e a)
An explicit parser from Text

-> IO (a, Text)

Like hGetSensitiveLine but with default handles

hGetSensitiveLine :: Buildable e => (Handle, Handle) -> Text -> (Text -> Either e a) -> IO (a, Text)

Gather user inputs until a newline is met, hiding what's typed with a placeholder character.