Cardano.CLI

Description

Shared types and helpers for CLI parsing

CLI Execution

cli :: Mod CommandFields a -> ParserInfo a

Construct a CLI from a list of a commands

>>> runCli $ cli $ cmdA <> cmdB <> cmdC

runCli :: ParserInfo (IO ()) -> IO ()

Runs a specific command parser using appropriate preferences

Commands

cmdMnemonic :: Mod CommandFields (IO ())

cmdWallet :: forall t. (DecodeAddress t, EncodeAddress t) => Mod CommandFields (IO ())

cmdTransaction :: forall t. (DecodeAddress t, EncodeAddress t) => Mod CommandFields (IO ())

cardano-wallet transaction

cmdAddress :: forall t. (DecodeAddress t, EncodeAddress t) => Mod CommandFields (IO ())

cmdStakePool :: forall t. (DecodeAddress t, EncodeAddress t) => Mod CommandFields (IO ())

cmdNetwork :: forall t. (DecodeAddress t, EncodeAddress t) => Mod CommandFields (IO ())

cmdVersion :: Mod CommandFields (IO ())

cmdKey :: Mod CommandFields (IO ())

Option & Argument Parsers

optionT :: FromText a => Mod OptionFields a -> Parser a

Helper for writing an option Parser using a FromText instance.

argumentT :: FromText a => Mod ArgumentFields a -> Parser a

Helper for writing an argument Parser using a FromText instance.

databaseOption :: Parser FilePath

-database=DIR

hostPreferenceOption :: Parser HostPreference

--listen-address=HOSTSPEC

, default: 127.0.0.1

listenOption :: Parser Listen

--random-port|--port=INT

nodePortOption :: Parser (Port "Node")

--node-port=INT

, default: 8080

nodePortMaybeOption :: Parser (Maybe (Port "Node"))

--node-port=INT

, default: use any available port

stateDirOption :: FilePath -> Parser (Maybe FilePath)

-state-dir=DIR, default: ~.cardano-wallet$backend/$network

syncToleranceOption :: Parser SyncTolerance

-sync-tolerance=DURATION, default: 300s

data LoggingOptions tracers

Constructors

LoggingOptions

Fields

loggingMinSeverity :: Severity

loggingTracers :: tracers

loggingTracersDoc :: Maybe Void

helperTracing :: [(String, String)] -> Parser (a -> a)

A hidden "helper" option which always fails, but shows info about the logging options.

loggingOptions :: Parser tracers -> Parser (LoggingOptions tracers)

loggingSeverities :: [(String, Severity)]

loggingSeverityOrOffReader :: ReadM (Maybe Severity)

loggingSeverityReader :: ReadM Severity

Types

data Service

Wrapper type around Text to make its semantic more explicit

data TxId

data MnemonicSize

Represents the number of words in a mnemonic sentence.

Only valid sizes are representable by this type.

Constructors

MS_9

MS_12

MS_15

MS_18

MS_21

MS_24

newtype Port (tag :: Symbol)

Port number with a tag for describing what it is used for

Constructors

Port

Fields

getPort :: Int

data CliKeyScheme key m

Record with mnemonic and key derivation funcionality — without any type parameters related to scheme.

This means that we can have a value for byron, a value for icarus, both with the same type CliKeyScheme.

CliKeyScheme is on the other hand parameterized over key and m. hoistKeyScheme is provided for mapping over m. mapKey is provided for mapping over key.

This way we can test mapping XPrvs to Text as pure code in Either String, rather than IO.

Constructors

CliKeyScheme

Fields

allowedWordLengths :: [Int]

mnemonicToRootKey :: [Text] -> m key

deriveChildKey :: key -> DerivationIndex -> m key

data CliWalletStyle

Constructors

Icarus

Ledger

Trezor

newtype DerivationIndex

Constructors

DerivationIndex

Fields

unDerivationIndex :: Word32

newtype DerivationPath

Constructors

DerivationPath [DerivationIndex]

newCliKeyScheme :: CliWalletStyle -> CliKeyScheme XPrv (Either String)

xPrvToTextTransform :: (XPrv -> Either String Text, Text -> Either String XPrv)

Pair of functions representing the bidirectional transformation between a XPrv and its 96-byte hex-encoded form.

hoistKeyScheme :: (forall a. m1 a -> m2 a) -> CliKeyScheme key m1 -> CliKeyScheme key m2

Change the underlying monad of a CliKeyScheme.

mapKey :: Monad m => (key1 -> m key2, key2 -> m key1) -> CliKeyScheme key1 m -> CliKeyScheme key2 m

Map over the key type of a CliKeyScheme.

deriveChildKey both takes a key as an argument and returns one. Therefore we need a bidirectional mapping between the two key types.

Logging

withLogging

Arguments

:: Maybe FilePath
Configuration file - uses default otherwise.

-> Severity
Minimum severity level to log

-> ((Configuration, Trace IO Text) -> IO a)
The action to run with logging configured.

-> IO a

Run an action with logging available and configured. When the action is finished (normally or otherwise), log messages are flushed.

ANSI Terminal Helpers

putErrLn :: Text -> IO ()

Like hPutErrLn but with provided default Handle

hPutErrLn :: Handle -> Text -> IO ()

Print an error message in red

enableWindowsANSI :: IO ()

The IOHK logging framework prints out ANSI colour codes with its messages. On Windows 10 and above it's possible to enable processing of these colour codes. The hSupportsANSIWithoutEmulation function does this as a side effect. On older versions of Windows, special treatment is required (see: ANSI). In this case, this function will achieve nothing, and the ANSI control characters will be printed in grey (too bad).

Working with Sensitive Data

getLine :: Buildable e => Text -> (Text -> Either e a) -> IO (a, Text)

Like hGetLine but with default handles

hGetLine :: Buildable e => (Handle, Handle) -> Text -> (Text -> Either e a) -> IO (a, Text)

Prompt user and parse the input. Re-prompt on invalid inputs.

getSensitiveLine

Arguments

:: Buildable e

=> Text
A message to prompt the user

-> (Text -> Either e a)
An explicit parser from Text

-> IO (a, Text)

Like hGetSensitiveLine but with default handles

hGetSensitiveLine :: Buildable e => (Handle, Handle) -> Text -> (Text -> Either e a) -> IO (a, Text)

Gather user inputs until a newline is met, hiding what's typed with a placeholder character.

Helpers

decodeError :: ByteString -> Maybe Text

Decode API error messages and extract the corresponding message.

requireFilePath :: FilePath -> IO ()

Look whether a particular filepath is correctly resolved on the filesystem. This makes for a better user experience when passing wrong filepaths via options or arguments, especially when they get forwarded to other services.

getDataDir

Arguments

:: String
The network backend name.

-> IO FilePath

Find the user data directory for a given node network backend.

setupDirectory :: (Text -> IO ()) -> FilePath -> IO ()

Initialize a directory to store data such as blocks or the wallet databases

waitForService

Arguments

:: Service
Name of the service

-> Tracer IO WaitForServiceLog
A Trace for logging

-> Port "node"
TCP Port of the service

-> IO ()
Service we're waiting after.

-> IO ()

Wait for a service to become available on a given TCP port. Exit on failure with a proper error message.

data WaitForServiceLog

Log messages from waitForService

Constructors

MsgServiceWaiting Service (Port "node")

MsgServiceReady Service

MsgServiceTimedOut Service

MsgServiceErrNetworkInvalid Text